Changes missed by CoreLib mirror
